+++ lib/widgets/dialogs/dialog_widget.dart
@@ -9,11 +9,7 @@ class DialogWidget extends StatelessWidget {
     this.title,
     this.msg,
     this.actions,
-    this.animation,
-    this.animationFrameRate,
-    this.animationRepeat,
-    this.animationAnimate,
-    this.animationReverse,
+    this.animationBuilder,
     this.customView,
     this.titleStyle,
     this.msgStyle,
@@ -21,40 +17,28 @@ class DialogWidget extends StatelessWidget {
   });
 
   /// [actions]Widgets to display a row of buttons after the [msg] widget.
-  final List<Widget> actions;
+  final List<Widget>? actions;
 
   /// [customView] a widget to display a custom widget instead of the animation view.
-  final Widget customView;
+  final Widget? customView;
 
   /// [title] your dialog title
-  final String title;
+  final String? title;
 
   /// [msg] your dialog description message
-  final String msg;
+  final String? msg;
 
-  /// [animation] lottie animations path
-  final String animation;
-
-  /// [frameRate] lottie animation framerate
-  final FrameRate animationFrameRate;
-
-  /// [repeat] should the lottie animation repeat?
-  final bool animationRepeat;
-
-  /// [animate] should the animation animate?
-  final bool animationAnimate;
-
-  /// [reverse] should the animation reverse?
-  final bool animationReverse;
+  /// [animationBuilder] lottie animations builder
+  final LottieBuilder? animationBuilder;
 
   /// [titleStyle] dialog title text style
-  final TextStyle titleStyle;
+  final TextStyle? titleStyle;
 
   /// [animation] lottie animations path
-  final TextStyle msgStyle;
+  final TextStyle? msgStyle;
 
   /// [color] dialog's backgorund color
-  final Color color;
+  final Color? color;
 
   @override
   Widget build(BuildContext context) {
@@ -62,29 +46,18 @@ class DialogWidget extends StatelessWidget {
       crossAxisAlignment: CrossAxisAlignment.center,
       mainAxisSize: MainAxisSize.min,
       children: [
-        animation != null
-            ? Container(
-                padding: EdgeInsets.only(top: 20),
-                height: 200,
-                width: double.infinity,
-                child: Lottie.asset(
-                  animation,
-                  animate: animationAnimate,
-                  frameRate: animationFrameRate,
-                  repeat: animationRepeat,
-                  reverse: animationReverse,
-                  fit: BoxFit.contain,
-                ),
-              )
-            : SizedBox(
-                height: 0,
-              ),
-        customView,
+        if (animationBuilder != null)
+          Container(
+            padding: EdgeInsets.only(top: 20),
+            height: 200,
+            width: double.infinity,
+            child: animationBuilder,
+          ),
         title != null
             ? Padding(
                 padding: const EdgeInsets.only(right: 20, left: 20, top: 24.0),
                 child: Text(
-                  title,
+                  title!,
                   style: titleStyle,
                 ),
               )
@@ -95,14 +68,14 @@ class DialogWidget extends StatelessWidget {
             ? Padding(
                 padding: const EdgeInsets.only(right: 20, left: 20, top: 16.0),
                 child: Text(
-                  msg,
+                  msg!,
                   style: msgStyle,
                 ),
               )
             : SizedBox(
                 height: 20,
               ),
-        actions != null && actions.isNotEmpty
+        actions?.isNotEmpty == true
             ? buttons(context)
             : SizedBox(
                 height: 20,
@@ -117,7 +90,7 @@ class DialogWidget extends StatelessWidget {
           const EdgeInsets.only(right: 20, left: 20, top: 16.0, bottom: 20.0),
       child: Row(
         mainAxisAlignment: MainAxisAlignment.center,
-        children: List.generate(actions.length, (index) {
+        children: List.generate(actions!.length, (index) {
           final TextDirection direction = Directionality.of(context);
           double padding = index != 0 ? 8 : 0;
           double rightPadding = 0;
@@ -128,7 +101,7 @@ class DialogWidget extends StatelessWidget {
           return Expanded(
             child: Padding(
               padding: EdgeInsets.only(left: leftPadding, right: rightPadding),
-              child: actions[index],
+              child: actions![index],
             ),
           );
         }),