+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.0.1]
+# Breakout changes:
+- remove `animation`, `frameRate`, `reverse` and replace it with `lottieBuilder` 
+# New features
+- add on close callback
+- adding custom view, in case you wanted to use only image or different size animations.
+
+
+## [1.0.0]
+- Migrate to Sound null safety
+- remove deprecated code
+
 ## [0.1.5]
 - Update dependencies
 
@@ -36,16 +48,20 @@ with a custom view
 - allow only titles dialogs
 - allow only buttons dialogs
 
-## [0.0.4] - adding RTL layouts support
+## [0.0.4] 
+- adding RTL layouts support
 
-## [0.0.3+1] - edit examples
+## [0.0.3+1] 
+- edit examples
 
-## [0.0.3] - breakout changes
-
+## [0.0.3] 
+- breakout changes
 - use actions widgets list to add buttons.
 - add out of the box customized buttons widgets.
 
 
-## [0.0.2] - fix single button padding bug and edit ReadMe.
+## [0.0.2]
+- fix single button padding bug and edit ReadMe.
 
-## [0.0.1] - first release.
+## [0.0.1] 
+- first release.